feat: lazy load heavy mobile sections
