+++ src/categories.js
@@ -1,0 +1,4 @@
+export const CATEGORIES = [
+  '食費','住居・光熱','日用品・消耗品','通信','交通・移動','医療・健康','衣服・美容',
+  '趣味・娯楽','旅行・レジャー','教育・書籍','交際費','ビジネス','税金・保険','その他','収入'
+];
